+++ netbox_certificates/migrations/0018_service_tags_related_name.py
@@ -17,6 +17,7 @@ class Migration(migrations.Migration):
                     name="tags",
                     field=taggit.managers.TaggableManager(
                         through="extras.TaggedItem",
+                        to="extras.Tag",
                         ordering=("weight", "name"),
                         manager=extras.managers.NetBoxTaggableManager,
                         related_name="netbox_certificates_service_tagged+",